speedup attach_label() function
"properties" paremeter is now an int instead of a gdouble

No need to convert from int to gdouble and then back from gdouble to
int.

+++ src/etats_csv.c
@@ -39,7 +39,7 @@
 
 /*START_STATIC*/
 static void csv_attach_hsep ( gint x, gint x2, gint y, gint y2);
-static void csv_attach_label ( gchar * text, gdouble properties, gint x, gint x2, gint y, gint y2,
+static void csv_attach_label ( gchar * text, int properties, gint x, gint x2, gint y, gint y2,
 							  GtkJustification align, gint transaction_number, const gchar *class );
 static void csv_attach_vsep ( gint x, gint x2, gint y, gint y2);
 static gint csv_finish ( void );
@@ -84,7 +84,7 @@ static gint csv_lastline = 1;
  * \param transaction_number the number a transaction to link to (not used as csv
  *            backend is not interactive)
  */
-void csv_attach_label ( gchar * text, gdouble properties, gint x, gint x2, gint y, gint y2,
+void csv_attach_label ( gchar * text, int properties, gint x, gint x2, gint y, gint y2,
                         GtkJustification align, gint transaction_number, const gchar *class )
 {
     gint pad;
